usm: http2: tests: Move StartH2CServer into a shared location
